디버깅으로 인자순서 바로잡음, 64비트 컴퓨터에서 sizeof(long) == sizeof(longlong) == 주소값 == 8, 32비트 컴퓨터에서 sizeof(long) == 주소값 == 4 이므로, long으로 받는 게 좋겠다.

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yehan <yehan@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 17:35:40 by yehan             #+#    #+#             */
-/*   Updated: 2022/03/28 11:39:49 by yehan            ###   ########.fr       */
+/*   Updated: 2022/03/28 13:38:25 by yehan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,8 +34,8 @@ typedef struct s_notation
 int		ft_printf(const char *format, ...);
 int		putarg_cnt(va_list *ap, char type);
 void	init_struct(t_notation *notation, char type);
-int		putnbr_type(long long nbr, char type);
-void	putnbr_long(unsigned long long nbr, t_notation notation, int *cnt);
+int		putnbr_type(long nbr, char type);
+void	putnbr_long(unsigned long nbr, t_notation notation, int *cnt);
 void	putnbr_int(unsigned int nbr, t_notation notation, int *cnt);
 
 #endif
